requirements.txt
app==0.0.1
fastapi==0.108.0
passlib==1.7.4
pydantic==2.5.3
pymongo==4.6.1
pytest==7.4.4
python-dotenv==1.0.0
python_jose==3.3.0
uvicorn==0.25.0
python-multipart